requirements.txt
alembic==1.13.2
blinker==1.8.2
cffi==1.17.0
click==8.1.7
colorama==0.4.6
cryptography==43.0.0
Flask==1.1.2
Flask-Migrate==2.6.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
Flask-WTF==1.2.1
greenlet==3.0.3
itsdangerous==2.0.1
Jinja2==2.10.2
Mako==1.3.5
MarkupSafe==2.0.1
psycopg2==2.9.9
pycparser==2.22
PyJWT==2.9.0
SQLAlchemy==2.0.32
typing_extensions==4.12.2
Werkzeug==2.0.3
WTForms==3.1.2
